pyproject.toml

[build-system]
requires = ["setuptools >= 64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "aind-smartspim-transform-utils"
description = "Transforms coordinates between SmartSPIM template and CCFv3"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "Allen Institute for Neural Dynamics"}
]
classifiers = [
    "Programming Language :: Python :: 3"
]
readme = "README.md"
dynamic = ["version"]

dependencies = [
    'ome-zarr==0.8.3',
    'zarr==2.18.0',
    'dask[distributed]==2024.4.1',
    'antspyx==0.4.2',
    'numpy==1.23.5',
    'boto3==1.37.2',
    'imlib==0.1.3'
]

[project.optional-dependencies]
dev = [
    'black',
    'coverage',
    'flake8',
    'interrogate',
    'isort',
    'Sphinx',
    'furo'
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "aind_smartspim_transform_utils.__version__"}

[tool.black]
line-length = 79
target_version = ['py36']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
  )/
  | .gitignore
)
'''

[tool.coverage.run]
omit = ["*__init__*"]
source = ["aind_smartspim_transform_utils"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "from",
    "import",
    "pragma: no cover"
]
fail_under = 100

[tool.isort]
line_length = 79
profile = "black"

[tool.interrogate]
ignore-init-method = true
ignore-private = true
exclude = ["setup.py", "docs", "build"]
ignore-regex = [
    "_check_path",
    "_fetch_zarr_data", 
    "__init__",
    "_load_data_from_s3",
    "_load_data_from_local", 
    "_read_xml_as_df",
    "load_cell_locations"
]
fail-under = 100